Switch to DualTreeBidiMap with custom comparator due to #3162
